+++ R/create_context.R
@@ -16,31 +16,33 @@
 #' @return A list with the context information for the learnitgrid Shiny app.
 #' @export
 create_context <- function(correction, base_corr_dir, base_templ_dir,
-base_repos_dir, repositories, assignments, github_url, branch) {
+    base_repos_dir, repositories, assignments, github_url, branch) {
+  # Note: for now, it is hardcoded in default_correction at the top of the file!
+  # Populate a list of correction folders available
   course_dirs <- dir_ls(base_corr_dir, type = "directory")
   if (!length(course_dirs))
-    stop("There is still no evalaution grids for this assignment. ",
-      "Use the script '04create_assessments4.R' to create them...")
+    stop("Il n'y a pas encore de grilles de correction disponibles. ",
+      "Utilisez le script '04create_assessments3.R' pour les cr\u00e9er...")
   courses <- basename(course_dirs)
 
   corr_dirs <- dir_ls(course_dirs, type = "directory")
   if (!length(corr_dirs))
-    stop("There is still no evalaution grids for this assignment. ",
-      "Use the script '04create_assessments4.R' to create them...")
+    stop("Il n'y a pas encore de grilles de correction disponibles. ",
+      "Utilisez le script '04create_assessments3.R' pour les cr\u00e9er...")
   corrections <- basename(corr_dirs)
 
   if (is.null(correction) || !length(correction) || correction == "")
-    stop("Correction set not provided, or operation cancelled.")
+    stop("Correction set non fourni, ou op\u00e9ration annul\u00e9e.")
   if (!correction %in% corrections)
-    stop("Correction set '", correction, "' not found. Check it...")
+    stop("Correction set '", correction, "' introuvable. V\u00e9rifiez sa valeur...")
 
   # Get corr_dir for this correction + get course, assignment and corr_date
   corr_dir <- corr_dirs[corrections == correction]
   course <- basename(dirname(corr_dir))
   corr_parts <- strsplit(correction, "_", fixed = TRUE)[[1]]
   if (length(corr_parts) < 4)
-    stop("Problem with the correction set '", correction,
-      "'  : it should be something like A00Ia_21M_title_2022-01-01")
+    stop("Il y a un probl\u00e8me avec le set de corrections '", correction,
+      "'  : il devrait \u00eatre quelque chose comme A00Ia_21M_titre_2022-01-01")
   corr_date  <- corr_parts[length(corr_parts)]
   assignment <- paste(corr_parts[-length(corr_parts)], collapse = "_")
   rm(corr_parts)
@@ -81,13 +83,14 @@ base_repos_dir, repositories, assignments, github_url, branch) {
   #if (!is.null(app))
   #  assign_infos <- assign_infos[tolower(assign_infos$app) == tolower(app), ]
   if (!NROW(assign_infos)) {
-    stop("The selected assignment (", assignment, ") is not found in ",
-      "the table.")
+    stop("L'exercice s\u00e9lectionn\u00e9 (", assignment,
+      ") n'est pas trouv\u00e9 dans la table des exercices.")
   }
 
   if (NROW(assign_infos) > 1) {
-    warning("The selected assignment (", assignment,
-      ") is found in several variants in the table. Use of the first one.")
+    warning("L'exercice s\u00e9lectionn\u00e9 (", assignment,
+      ") est trouv\u00e9 dans la table ",
+      " des exercices en plusieurs exemplaires. Utilisation du premier.")
     assign_infos <- assign_infos[1, ]
   }
 
@@ -102,7 +105,7 @@ base_repos_dir, repositories, assignments, github_url, branch) {
     templ_corrs[substring(templ_corrs$criterion, 1, 8) != "!{login}", ]
 
   # Path of git_stat.csv
-  git_stats <- file_path_check(repos_dir, "git_stats.csv")
+  git_dir <- file_path_check(repos_dir, "git_stats.csv")
 
   # The context object (list) contains info required to populate tables
   context <- list(
@@ -112,7 +115,7 @@ base_repos_dir, repositories, assignments, github_url, branch) {
     assign_infos = assign_infos, # Informations about the assignment
     github_url = github_url,     # The base URL for GitHub links
     branch = branch,             # The GitHub branch
-    git_stats_file = git_stats,  # The path to git_stats.csv file
+    git_dir = git_dir,           # The path to git_stats.csv
     user_logins = user_logins,   # The users logins (teams for group projects)
     repos_names = repos_names,   # The names of repositories
     repos_names2 = repos_names2, # The names of repos if different from apps