+++ R/get_git_stats.R
@@ -3,7 +3,7 @@
 #' @description Use data from a git_stats.csv file to get a history of commits
 #'   done in a repository for the learnitgrid Shiny app.
 #'
-#' @param git_stats_file The path the the git_stats.csv file
+#' @param dir The path the the git_stats.csv file
 #' @param exclude_authors The list of authors to exclude from the stats
 #' @param type The type of files to consider ("all", "R", "Rmd", or "Qmd")
 #' @param tz The time zone to use for times
@@ -11,15 +11,15 @@
 #' @return A data frame with git stats data to be used by the learnitgrid Shiny
 #'   app.
 #' @export
-get_git_stats <- function(git_stats_file,
-exclude_authors = "github-classroom[bot]", type = "all", tz = "UTC") {
+get_git_stats <- function(dir, exclude_authors = "github-classroom[bot]",
+    type = "all", tz = "UTC") {
 
-  if (!file_exists(git_stats_file))
-    stop("The file ", git_stats_file, " does not exist")
-  stat <- suppressMessages(read.csv(git_stats_file))
+  if (!file_exists(dir))
+    stop("The directory ", dir, " does not exist")
+  stat <- suppressMessages(read.csv(dir))
 
   # TODO: allow for more types too (e.g., doc, docx, pdf, ppt, pptx, ...)
-  vec <- match.arg(type, choices = c("all", "R", "Rmd", "Qmd"),
+  vec <- match.arg(type, choices = c("all", "R", "Rmd", "qmd"),
     several.ok = TRUE)
 
   if (length(exclude_authors))
@@ -41,3 +41,5 @@ exclude_authors = "github-classroom[bot]", type = "all", tz = "UTC") {
 
   stat
 }
+
+#res <- stat_git(git_stats, type = c("Rmd", "R"))