+++ R/learnitgrid-package.R
@@ -20,7 +20,7 @@
 #' @importFrom glue glue
 #' @importFrom fs path dir_create dir_exists dir_ls file_exists link_create link_delete link_exists
 #' @importFrom collapse fcumsum GRP roworder
-#' @importFrom ggplot2 geom_histogram facet_wrap geom_point geom_step geom_vline labs theme xlim
+#' @importFrom ggplot2 geom_histogram element_text facet_wrap geom_point geom_step geom_vline labs theme xlim
 #' @importFrom stats fivenum
 #' @importFrom lubridate ymd_hms with_tz
 #' @importFrom parsermd as_document by_section has_label has_type parse_rmd rmd_select
@@ -30,7 +30,7 @@
 #' @importFrom shinydashboard dashboardBody dashboardHeader dashboardPage dashboardSidebar
 #' @importFrom shinycssloaders withSpinner
 #' @importFrom stringdist stringsim
-#' @importFrom utils read.csv write.csv
+#' @importFrom utils browseURL read.csv untar write.csv
 #' @importFrom writexl write_xlsx
 # The following block is used by usethis to automatically manage
 # roxygen namespace tags. Modify with care!